feat: Add support for default desktop icon in InnoSetup
This commit adds support for a default desktop icon in InnoSetup. The `InnoSetupDesktopIconBuilder` class now accepts a `defaultDesktopIcon` parameter, which determines whether the desktop icon should be checked by default. The `InnoSetup` class has also been updated to pass the `defaultDesktopIcon` value to the `InnoSetupDesktopIconBuilder` when creating the desktop icon task.